refactor(appearance): share canonical atlas sampling orchestration (#4381)
